+++ tweaks/ytmusic/src/Prefs/YTMUltimateSettingsController.h
@@ -1,0 +1,10 @@
+#import <UIKit/UIKit.h>
+#import "PlayerSettingsController.h"
+#import "ThemeSettingsController.h"
+#import "NavBarSettingsController.h"
+#import "TabBarSettingsController.h"
+#import "TranslationSettingsController.h"
+
+@interface YTMUltimateSettingsController : UIViewController <UITableViewDelegate, UITableViewDataSource>
+@property (nonatomic, strong) UITableView* tableView;
+@end